Rename analyzers
