+++ analyze/__init__.py
@@ -1,12 +1,12 @@
-from .checkerflickeranalyzer import checkerflicker
-from .checkerflickerplusanalyzer import checkerflickerplus
-from .fffanalyzer import fff
-from .OMBanalyzer import omb
-from .OMSpatchesanalyzer import omspatches
-from .onoffstepsanalyzer import onoffsteps
-from .saccadegratingsanalyzer import saccadegratings
-from .spontanalyzer import spont
-from .stripeflickeranalysis import stripeflicker
+from .checkerflicker import checkerflicker
+from .checkerflickerplus import checkerflickerplus
+from .fff import fff
+from .omb import omb
+from .omspatches import omspatches
+from .onoffsteps import onoffsteps
+from .saccadegratings import saccadegratings
+from .spont import spont
+from .stripeflicker import stripeflicker
 from .stripesurround import stripesurround
 
 from .allfff import allfff
@@ -14,17 +14,17 @@ from .allonoff import allonoff
 from .experiment import experiment
 
 __all__ = [
-        'experiment',
-        'checkerflicker',
-        'checkerflickerplus',
-        'fff',
-        'omb',
-        'omspatches',
-        'onoffsteps',
-        'saccadegratings',
-        'spont',
-        'stripeflicker',
-        'stripesurround',
-        'allfff',
-        'allonoff',
-        ]
+    'experiment',
+    'checkerflicker',
+    'checkerflickerplus',
+    'fff',
+    'omb',
+    'omspatches',
+    'onoffsteps',
+    'saccadegratings',
+    'spont',
+    'stripeflicker',
+    'stripesurround',
+    'allfff',
+    'allonoff',
+]
