+++ analyze/checkerflicker.py
