+++ analyze/omspatches.py
