+++ analyze/spont.py
