+++ analyze/stripeflicker.py
